requirements.txt
awsebcli==3.10.4
blessed==1.14.2
botocore==1.5.87
cement==2.8.2
colorama==0.3.7
Django==1.11.3
django-filter==1.0.4
django-localflavor==1.5.2
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.3
Markdown==2.6.8
pathspec==0.5.0
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.44.0
